Reduce GELU parity for testing
-Reduce hidden layer to 1
-Reduce testing to only GPU
-Perform 2 test cases instead of 100

With this config we're able to see a failure always on the second kernel exectution.
There seems to be some interplay with calls that ORT does into HIP that breaks
how and when we call hipMemcpy(). We Launch the second kernel and consistently
break accuracy as the hipMemCpy(DeviceToHost) is performed before the actual kernel
has completed the run.